+++ packages/ui/src/components/Select/index.ts
@@ -1,1 +1,1 @@
-export * from "./Select"
+export * from "./MobileSelect"
